Add poly-int.h
This patch adds a new class to represent polynomail integers
of the form:

  C0 + C1*X1 + C2*X2 ... + Cn*Xn

It also adds typedefs for polynomial integers that represent
offsets and sizes of various precisions.  In these typedefs, the Ci
coefficients are compile-time constants and the Xi indeterminates
are run-time invariants.  The number of coefficients is controlled
by the target and is initially 1 for all ports.

The implementation of some routines only handle the cases N==1
and N==2.  Support for other coefficient counts can be added
when needed.  Since the core arithmetic operations are heavily
used, and since the routines just mentioned don't support N>2,
the file unrolls the implementation of these core operations
and makes them ALWAYS_INLINE, to reduce the compile-time impact
when GCC is built at -O0.

The patch also adds a new macro, TARGET_C_FILE, that can be
set to indicate that a TU contains target-specific rather
than target-independent code.  When this macro is set and
the number of coefficients is 1, the polynomial integer
class defines a conversion operator to a constant.  This
allows most existing target code to work without modification.
The only exception is usually in ?: expressions, in cases
where one arm ends up being a polynomial and the other
remains a constant.  In these cases it would be valid to
convert the constant to a polynomial and the polynomial
to a constant, so the type is ambiguous and the expression
ill-formed.  Casts are added by later patches as necessary.

The patch also adds a new target hook to return the estimated
value of a polynomial integer for costing purposes.

2017-09-04  Richard Sandiford  <richard.sandiford@linaro.org>
	    Alan Hayward  <alan.hayward@arm.com>
	    David Sherwood  <david.sherwood@arm.com>

gcc/
	* poly-int.h: New file.
	* poly-int-types.h: Likewise.
	* coretypes.h: Include them.
	(POLY_INT_CONVERSION): Define.
	* target.def (estimated_poly_value): New hook.
	* doc/tm.texi.in (TARGET_ESTIMATED_POLY_VALUE): New hook.
	* doc/tm.texi: Regenerate.
	* doc/poly-int.texi: New file.
	* doc/gccint.texi: Include it.
	* Makefile.in (TEXI_GCCINT_FILES): Add poly-int.texi.
	* genmodes.c (NUM_POLY_INT_COEFFS): Provide default definition.
	(emit_insn_modes_h): Emit a definition of NUM_POLY_INT_COEFFS.
	* targhooks.h (default_estimated_poly_value): Declare.
	* targhooks.c (default_estimated_poly_value): New function.
	* target.h (estimated_poly_value): Likewise.
	* wide-int.h (WI_UNARY_RESULT): Use wi::binary_traits.
	(wi::unary_traits): Delete.
	(operator /): New function.
	(operator %): Likewise.