[HSABE] hsa_kernel and hsa_function function attributes
This patch is mostly the "direct HSA" patch set of Martin Jambor.

Martin's patch adds new function attributes 'hsa_kernel' and
'hsa_function' for marking C/C++ functions for HSA offloading.
'hsa_kernel' makes the function a host-callable kernel while
'hsa_function' makes the function compiled to an HSA-only program
scope helper function.  It also introduces a list of builtins used
to call special HSAIL instructions.

This patch extends the original patch slightly with a feature to
specialize BRIGs by allowing redefining function declarations to
ones determined at finalization time.  This enables avoiding
indirect calls in kernels in cases where the function call targets
are known at the invoke time.  Another option would be to
use 'icall' but it is supported only in the HSA Full profile and
it would then always leave unnecessary indirect calls to the kernels which
hinders optimizations (esp. inlining and vectorization) and can ruin
performance on GPUs.

This was implemented via additional parameters to the invoke function
which allow listing host-agent function pairs: The host function names
are used to find the corresponding Agent-versions in the function
index, and their Agent side addresses are resolved and _patched to
the BRIG_ before passing it to the finalizer.